Create rsync_interactive.py
Python script, when run the script asks the user a series of questions.  The user's answers are then appended as options onto the 'rsync' command.  Additionally, The user is asked to input the 'source' & 'destination' directories, which are appended as well. Next, the script initiates the command as the user intended. Finally, after the 'rsync' command finishes, a log file is created in the destination directory, if the log file already exists, then the data is appended to it.